Moves date range helpers from dashboard helper to daterange helper. Adds additional class that autosubmits the filter on change. Renames dashboard filter to use universal filter name
